# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blocker X - Discord Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invite/redeem:
    post:
      operationId: redeemInviteCode
      tags: [invite]
      summary: Redeem an invitation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteCodeInput"
      responses:
        "200":
          description: Code redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Profile
...
  /files/{botId}/list:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files for a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: dirPath
          in: query
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: Files list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileEntry"
  /files/{botId}/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
            default: "/"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileEntry"
  /files/{botId}/delete:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or folder
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilePathInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /files/{botId}/rename:
    patch:
      operationId: renameFile
      tags: [files]
      summary: Rename a file
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRenameInput"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /files/{botId}/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file contents
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: filePath
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /files/{botId}/write:
    put:
      operationId: writeFile
      tags: [files]
      summary: Write/update file contents
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWriteInput"
      responses:
        "200":
          description: File saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /files/{botId}/mkdir:
    post:
      operationId: createFolder
      tags: [files]
      summary: Create a folder
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilePathInput"
      responses:
        "200":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Environment Variables
  /env/{botId}:
    get:
      operationId: listEnvVars
      tags: [env]
      summary: List environment variables for a bot
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Env vars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnvVar"

    post:
      operationId: setEnvVar
      tags: [env]
      summary: Create or update an environment variable
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarInput"
      responses:
        "200":
          description: Env var saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVar"
  /env/{botId}/{varId}:
    delete:
      operationId: deleteEnvVar
      tags: [env]
      summary: Delete an environment variable
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: varId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Logs
  /logs/{botId}:
    get:
      operationId: getBotLogs
      tags: [logs]
      summary: Get bot logs
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
        - name: logLimit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  # AI Assistant
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    InviteCodeInput:
      type: object
      properties:
        code:
          type: string
      required:
        - code
...
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: ["integer", "null"]
        lastModified:
          type: ["string", "null"]
      required:
        - name
        - path
        - type
    FilePathInput:
      type: object
      properties:
        path:
          type: string
      required:
        - path
    FileRenameInput:
      type: object
      properties:
        oldPath:
          type: string
        newPath:
          type: string
      required:
        - oldPath
        - newPath
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
      required:
        - path
        - content
    FileWriteInput:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    FileUploadInput:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
        content:
          type: string
        encoding:
          type: string
      required:
        - path
        - name
        - content
    EnvVar:
      type: object
      properties:
        id:
          type: string
        botId:
          type: string
        key:
          type: string
        value:
          type: string
        createdAt:
          type: string
      required:
        - id
        - botId
        - key
        - value
        - createdAt
    EnvVarInput:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
    LogEntry:
      type: object
      properties:
        id:
          type: string
        botId:
          type: string
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        timestamp:
          type: string
      required:
        - id
        - botId
        - level
        - message
        - timestamp
...